requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.16.1
django-filter==24.1
drf-yasg==1.21.15
inflection==0.5.1
packaging==26.0
pytz==2026.1.post1
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
pandas>=2.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
tqdm>=4.66
matplotlib>=3.7
seaborn>=0.13
Pillow>=10.0
djangorestframework-simplejwt>=5.3.2
geopy>=2.4.0
python-dotenv>=1.0.0
